+++ RELEASE_NOTES.md
@@ -12,8 +12,9 @@
 - Fixes Find input error not going away after using Go To Next Find Match
 - Fixes Back and Forward to not leave inaccessible paths on stack
 - Fixes drive not being selected when opening parent
+- Fixes refresh error to replace item list with error message
 - Fixes Refresh window on focus with recursive search to refresh current directory
-- Undo/redo now reports status for each iteration so they can be seen in history
+- Undo/redo now reports status for each iteration so they can be seen in status history
 - Various optimizations for file write operations
 
 ### 0.6.2  (2022-12-23)